refactor(calendar): extract navigation buttons into separate component
